+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-05T14:12:53.503Z",
+    "2026-02-05T16:13:04.285Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {